requirements.txt
torch
torchvision
matplotlib
Pillow
numpy
scikit-image
scikit-learn